# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Blog API",
        "contact": {
            "email": "halkawt.abdulilah@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/activities": {
            "get": {
                "tags": [
                    "Activity Logs"
                ],
                "summary": "List activity logs with filters",
                "description": "Returns paginated activity logs with action, entity type, and actor filtering",
                "operationId": "3bf5edadbd6ce763463853e8f4d0ea4f",
                "parameters": [
                    {
                        "name": "action",
                        "in": "query",
                        "description": "Filter by action type",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "entity_type",
                        "in": "query",
                        "description": "Filter by entity type (e.g., 'post', 'category')",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "actor",
                        "in": "query",
                        "description": "Filter by actor name or ID",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "orderBy",
                        "in": "query",
                        "description": "Field to sort by",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "created_at",
                            "enum": [
                                "created_at",
                                "updated_at"
                            ]
                        }
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "description": "Sort direction",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "properties": {
                                                "page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "pages": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "limit": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 50
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ActivityLog"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logs/revert/{activityLog}": {
            "post": {
                "tags": [
                    "Activity Logs"
                ],
                "summary": "Revert an entity to its previous state using an activity log",
                "description": "Restores an entity's fields to their 'before' state from the specified activity log entry",
                "operationId": "18312510bf4397c6c460ba8c90cbc5af",
                "parameters": [
                    {
                        "name": "activityLog",
                        "in": "path",
                        "description": "ID of the activity log entry to revert from",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful revert",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully reverted to previous version"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Post"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Activity log or entity not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Activity log not found"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\ActivityLog] 1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Revert failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Revert failed"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Trying to get property 'before' of non-object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ActivityLog": {
                "description": "Activity log entry",
                "required": [
                    "id",
                    "action",
                    "entity_type",
                    "entity_id",
                    "actor",
                    "created_at"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier for the log entry",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "action": {
                        "description": "Action performed (e.g., created, updated, deleted)",
                        "type": "string",
                        "example": "updated"
                    },
                    "entity_type": {
                        "description": "Type of the affected entity",
                        "type": "string",
                        "example": "App\\\\Models\\\\Post"
                    },
                    "entity_id": {
                        "description": "ID of the affected entity",
                        "type": "integer",
                        "example": 42
                    },
                    "changed_fields": {
                        "description": "JSON object showing before/after values of changed fields",
                        "type": "object",
                        "example": {
                            "title": {
                                "before": "Old Title",
                                "after": "New Title"
                            },
                            "content": {
                                "before": "Old content",
                                "after": "Updated content"
                            }
                        },
                        "nullable": true
                    },
                    "actor": {
                        "description": "User or system that performed the action",
                        "type": "string",
                        "example": "john"
                    },
                    "created_at": {
                        "description": "When the action was logged",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-05-15T12:34:56Z"
                    },
                    "updated_at": {
                        "description": "When the log entry was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-05-15T12:34:56Z"
                    }
                },
                "type": "object"
            },
...
            "Post": {
                "required": [
                    "title",
                    "content",
                    "author",
                    "category_id"
                ],
                "properties": {
                    "id": {
                        "description": "Auto-incremented post ID",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "Post title",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Getting Started with Laravel"
                    },
                    "content": {
                        "description": "Main post content (HTML/markdown)",
                        "type": "string",
                        "example": "This is a detailed guide about Laravel..."
                    },
                    "author": {
                        "description": "Author display name",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Jane Doe"
                    },
                    "category_id": {
                        "description": "ID of the associated category",
                        "type": "integer",
                        "example": 2
                    },
                    "created_at": {
                        "description": "Record creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-05-10T08:00:00Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-05-12T09:15:00Z"
                    },
                    "deleted_at": {
                        "description": "Soft deletion timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    },
...
}